adjust the aspect ratio
